transcend the lower limit of hspec version (for debian wheezy hspec-1.1.0). remove dependency to hspec-discover (obsolete package).
