+++ simple-sendfile.cabal
@@ -49,13 +49,13 @@ Test-Suite spec
   Hs-Source-Dirs:       test
   Main-Is:              Spec.hs
   GHC-Options:          -Wall
-  Other-Modules:        SendfileSpec
   Build-Depends:        base
                       , HUnit
                       , bytestring
                       , conduit
                       , directory
-                      , hspec >= 1.3
+                      , hspec >= 1.1
+                      , hspec-expectations
                       , network
                       , network-conduit
                       , process
